feat(search)!: limit the search only in visible line
note:
default behavior is search to the end of the file, which is not
effective because vim-sneak doesn't give label if it beyond the visible
line in current window. that's why limiting the search only in visible
line in current window makes more sense.

if the pattern not in visible line in current window, then mark it as
`not found`.